src/requirements.txt
Flask==2.1.3
waitress==2.1.2
Pillow==9.2.0
urllib3==1.26.11
torch==1.8.1
paddlepaddle==2.3.1
paddleocr==2.5.0.3
